공유 UI 구현을 위한 PostCell 리팩토링
