+++ Facebook/Facebook/Features/PostDetail/Views/SubPostCell.swift
@@ -9,85 +9,12 @@ import UIKit
 import Kingfisher
 import SnapKit
 
-class SubPostCell: PostCell {
-    
+class SubPostCell: PostCell<SubPostContentView> {
+
     override class var reuseIdentifier: String { "SubPostCell" }
     
-    var singleImageView: UIImageView = {
-        let view = UIImageView()
-        view.contentMode = .scaleAspectFit
-        return view
-    }()
-    
-    func configureCell(with subPost: Post) {
-        self.post = subPost
-        textContentLabel.text = subPost.content
-        setImage(from: URL(string: subPost.file ?? ""))
-        let isEmptyContent = subPost.content.isEmpty
-        textContentLabel.isHidden = isEmptyContent
-        statHorizontalStackView.snp.remakeConstraints { make in
-            make.top.equalTo(isEmptyContent ? singleImageView.snp.bottom : textContentLabel.snp.bottom).offset(CGFloat.standardTopMargin)
-            make.leading.trailing.equalTo(contentView).inset(CGFloat.standardLeadingMargin)
-        }
+    override func configure(with newPost: Post, showGrid: Bool = false) {
+        postContentView.configure(with: newPost)
     }
     
-    override func setLayout() {
-        contentView.addSubview(singleImageView)
-        singleImageView.snp.makeConstraints { make in
-            make.leading.trailing.top.equalTo(contentView)
-            make.height.equalTo(300)  // default estimated height
-        }
-        
-        contentView.addSubview(textContentLabel)
-        textContentLabel.snp.makeConstraints { make in
-            make.top.equalTo(singleImageView.snp.bottom).offset(CGFloat.standardTopMargin)
-            make.leading.trailing.equalTo(contentView).inset(CGFloat.standardLeadingMargin)
-        }
-        
-        contentView.addSubview(statHorizontalStackView)
-        statHorizontalStackView.snp.makeConstraints { make in
-            make.top.equalTo(textContentLabel.snp.bottom).offset(CGFloat.standardTopMargin)
-            make.leading.trailing.equalTo(contentView).inset(CGFloat.standardLeadingMargin)
-        }
-        
-        contentView.addSubview(buttonHorizontalStackView)
-        buttonHorizontalStackView.snp.makeConstraints { make in
-            make.top.equalTo(statHorizontalStackView.snp.bottom).offset(CGFloat.standardTopMargin)
-            make.leading.trailing.equalTo(contentView).inset(CGFloat.standardLeadingMargin)
-            make.height.equalTo(CGFloat.buttonGroupHeight)
-        }
-        
-        contentView.addSubview(divider)
-        divider.snp.makeConstraints { make in
-            make.top.equalTo(buttonHorizontalStackView.snp.bottom)
-            make.leading.trailing.equalTo(contentView)
-            make.bottom.equalTo(contentView).priority(.high)
-            make.height.equalTo(5)
-        }
-    }
-    
-    func setImage(from url: URL?) {
-        guard let url = url else {
-            return
-        }
-        KF.url(url)
-            .setProcessor(KFProcessors.shared.downsampling)
-            .loadDiskFileSynchronously()
-            .cacheMemoryOnly()
-            .onSuccess { result in
-                self.singleImageView.snp.updateConstraints { make in
-                    make.height.equalTo(self.calcImageHeight(imageSize: result.image.size, viewWidth: self.frame.width))
-                }
-                self.layoutIfNeeded()
-            }
-            .onFailure { error in print("로딩 실패", error)}
-            .set(to: self.singleImageView)
-    }
-    
-    func calcImageHeight(imageSize: CGSize, viewWidth: CGFloat) -> CGFloat {
-        let myImageWidth = imageSize.width
-        let myImageHeight = imageSize.height
-        let ratio = viewWidth / myImageWidth
-        return myImageHeight * ratio
-    }
 }
